requirements.txt
streamlit
python-dotenv
pypdf
langchain
langchain-openai
langchain-community
chromadb
tiktoken
pydantic
python-docx
